requirements.txt
pydantic==2.8.2
pytest==8.3.1
python-dotenv==1.0.1
PyYAML==6.0.1
pipreqs
